First release. v1.0 22 Mar 2016
