+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "23.0.2"
 
     defaultConfig {
-        applicationId "com.example.rajat.samplefab"
+        applicationId "rajat.birthdaycardprakhar"
         minSdkVersion 15
         targetSdkVersion 23
         versionCode 1
