+++ app/src/main/java/com/example/rajat/birthdaycardprakhar/onFetchCompletion.java
@@ -1,0 +1,8 @@
+package com.example.rajat.birthdaycardprakhar;
+
+/**
+ * Created by hscuser on 21/03/16.
+ */
+public interface onFetchCompletion {
+    void onMsgFetchComplete();
+}
