--- app/src/main/java/com/example/rajat/samplefab/onFetchCompletion.java
@@ -1,8 +1,0 @@
-package com.example.rajat.samplefab;
-
-/**
- * Created by hscuser on 21/03/16.
- */
-public interface onFetchCompletion {
-    void onMsgFetchComplete();
-}
